+++ contracts/utils/MerkleProof.sol
@@ -1,19 +1,21 @@
 // SPDX-License-Identifier: MIT
 pragma solidity >=0.8.0 <0.9.0;
 
-import {SafeMath} from '@openzeppelin/contracts/utils/math/SafeMath.sol';
+import {SafeMath} from "@openzeppelin/contracts/utils/math/SafeMath.sol";
 
-library  MerkleProof {
+
+library MerkleProof {
     using SafeMath for uint32;
 
-    function verify (
-        bytes32 _leaf, bytes32 _root,  bytes32[] memory _proof, uint32 _index
-    )
-        internal pure returns (bool)
-    {
+    function verify(
+        bytes32 _leaf,
+        bytes32 _root,
+        bytes32[] memory _proof,
+        uint32 _index
+    ) internal pure returns (bool) {
         bytes32 hash = _leaf;
 
-        for (uint i = 0; i < _proof.length; i++) {
+        for (uint256 i = 0; i < _proof.length; i++) {
             bytes32 proofElement = _proof[i];
 
             if (_index.mod(2) == 0) {
